separate theme into helpers and preload helpers
