+++ app.js
@@ -1,17 +1,13 @@
 // Modules to control application life and create native browser window
-const {
-  BrowserWindow,
-  webContents,
-  ipcMain,
-  nativeTheme,
-  app,
-} = require("electron");
+const { BrowserWindow, ipcMain, nativeTheme, app } = require("electron");
 
 const store = require("./src/helpers/store");
 const config = require("./src/helpers/config");
+const { setThemeOnAllWindows } = require("./src/helpers/theme");
 const {
   CONSTANTS: { OS_PLATFORMS, THEME_OPTIONS },
 } = require("./src/helpers/util");
+
 const { createMainWindow } = require("./src/js/mainwindow");
 
 // Keep a global reference of the window object, if you don't, the window will
@@ -19,22 +15,9 @@ const { createMainWindow } = require("./src/js/mainwindow");
 let win, view;
 let childwin, childview;
 
-// Select which theme to change to
-const selectTheme = () => {
-  const userTheme = store.get("theme");
-  const OSTheme = nativeTheme.shouldUseDarkColors
-    ? THEME_OPTIONS.DARK
-    : THEME_OPTIONS.LIGHT;
-
-  // If theme is auto, select os theme.
-  // Else theme is manually selected, choose user's selection
-  return userTheme === THEME_OPTIONS.AUTO ? OSTheme : userTheme;
-};
-
 // Listen for theme requests from windows to set theme
 ipcMain.on("theme-request", function (_, webContentsId) {
-  const toThemeStyle = selectTheme();
-  webContents.fromId(webContentsId).send("theme-reply", toThemeStyle);
+  setThemeOnAllWindows();
 });
 
 // Listen for changes in native os theme to set theme
@@ -43,18 +26,12 @@ nativeTheme.on("updated", () => {
   if (store.get("theme") !== THEME_OPTIONS.AUTO) {
     return;
   }
-  const toThemeStyle = selectTheme();
-  // Send to all webcontents at once. This triggers an appwide theme change.
-  const allWebContents = webContents.getAllWebContents();
-  Promise.all(allWebContents.map((wc) => wc.send("theme-reply", toThemeStyle)));
+  setThemeOnAllWindows();
 });
 
 // Listen for changes in store
-store.onDidChange("theme", () => {  
-  const toThemeStyle = selectTheme();
-  // Send to all webcontents at once. This triggers an appwide theme change.
-  const allWebContents = webContents.getAllWebContents();
-  Promise.all(allWebContents.map((wc) => wc.send("theme-reply", toThemeStyle)));
+const unsubscribeStoreWatch = store.onDidChange("theme", () => {
+  setThemeOnAllWindows();
 });
 
 // This method will be called when Electron has finished
@@ -75,6 +52,8 @@ app.on("window-all-closed", function () {
   // On macOS it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   if (config.osPlatform !== OS_PLATFORMS.MAC_OS) app.quit();
+
+  unsubscribeStoreWatch();
 });
 
 // Make store accessible for easy access in other files.
